added - ajax class
